+++ google-cloud/fedora-coreos/kubernetes/variables.tf
@@ -60,7 +60,7 @@ variable "os_stream" {
 variable "disk_size" {
   type        = number
   description = "Size of the disk in GB"
-  default     = 40
+  default     = 30
 }
 
 variable "worker_preemptible" {