+++ google-cloud/flatcar-linux/kubernetes/workers/variables.tf
@@ -42,7 +42,7 @@ variable "os_image" {
 variable "disk_size" {
   type        = number
   description = "Size of the disk in GB"
-  default     = 40
+  default     = 30
 }
 
 variable "preemptible" {